planner classes

+++ trunk/ 3630-sp09-rovio/src/Waypoint.java
@@ -27,9 +27,9 @@ public class Waypoint {
 	}
 	
 	public double distance(double x, double y) {
-		int deltaX = x - this.x;
-		int deltaY = y - this.y;
-		int c2 = deltaX * deltaX + deltaY + deltaY;
+		double deltaX = x - this.x;
+		double deltaY = y - this.y;
+		double c2 = deltaX * deltaX + deltaY + deltaY;
 		return Math.sqrt(c2);
 	}
